feat: TaskGenerator uses filtered task view - Add state for no completeable tasks in filtered view - Extract filters into separate files and add useFilterTasks hook
