+++ src/components/TaskGenerator.js
@@ -2,39 +2,53 @@ import React, { useState, useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import useBreakpoint, { MEDIA_QUERIES, MODE } from '../hooks/useBreakpoint';
 import { toggleCompleted, updateRandomTask } from '../store/tasks/tasks';
-import tasks from '../data/tasks';
+import useFilterTasks from '../hooks/useFilterTasks';
 import SkillRequirementList from './SkillRequirementList';
 import Category from './Category';
 import Difficulty from './Difficulty';
+import TASKS from '../data/tasks';
 
 export default function TaskGenerator() {
     const [activeTask, setActiveTask] = useState(undefined);
-    const [allTasksCompleted, setAllTasksCompleted] = useState(false);
-
-    const { randomTaskId, tasks: reduxTasks } = useSelector(state => state.tasks);
-    const dispatch = useDispatch();
+    const [noPossibleTasks, setNoPossibleTasks] = useState(false);
     const isLgViewport = useBreakpoint(MEDIA_QUERIES.LG, MODE.GREATER);
-    const allTasks = Object.values(tasks);
+
+    const {
+        tasks: { tasks: tasksState, randomTaskId },
+        filters: { tasks: filterState },
+    } = useSelector(({ tasks, filters }) => ({ tasks, filters }));
+    const dispatch = useDispatch();
+
+    const filteredTasks = useFilterTasks({
+        filterState: {
+            ...filterState,
+        },
+        tasksState,
+    });
+
+    const allTasksCompleted =
+        Object.values(tasksState).filter(task => task.completed).length === Object.values(TASKS).length;
 
     useEffect(() => {
         if (randomTaskId) {
-            const storedTask = allTasks.find(task => task.id === randomTaskId);
+            const storedTask = Object.values(TASKS).find(task => task.id === randomTaskId);
             setActiveTask(storedTask);
         }
     }, []);
 
     const generateTask = () => {
-        const notPossibleTaskIDsFromStore = Object.entries(reduxTasks)
-            .filter(([, { completed, ignored }]) => completed || ignored)
-            .map(([taskId]) => taskId);
+        const completeTasksInState = Object.entries(tasksState)
+            .filter(([, taskInfo]) => taskInfo.completed)
+            .map(([id]) => id);
 
-        const possibleTasks = allTasks.filter(task => !notPossibleTaskIDsFromStore.includes(task.id));
-        const randomTask = possibleTasks[Math.floor(Math.random() * possibleTasks.length)];
+        const incompleteTasks = filteredTasks.filter(task => !completeTasksInState.includes(task.id));
+        const randomTask = incompleteTasks[Math.floor(Math.random() * filteredTasks.length)];
 
-        if (randomTask) setActiveTask(randomTask);
-        else setAllTasksCompleted(true);
+        setActiveTask(randomTask);
+        if (randomTask) setNoPossibleTasks(false);
+        else setNoPossibleTasks(true);
 
-        dispatch(updateRandomTask(randomTask.id));
+        dispatch(updateRandomTask(randomTask?.id || null));
     };
 
     const completeTask = taskId => {
@@ -43,7 +57,13 @@ export default function TaskGenerator() {
         setActiveTask(undefined);
     };
 
-    const renderCompleted = <p className='italic '>You have completed all the tasks! 🎉</p>;
+    const renderNoPossibleTask = (
+        <p className='italic '>
+            {allTasksCompleted
+                ? 'You have completed all tasks! 🎉'
+                : 'No incomplete tasks in table, please update your filters!'}
+        </p>
+    );
 
     const renderTask = activeTask ? (
         <>
@@ -60,42 +80,40 @@ export default function TaskGenerator() {
             </div>
         </>
     ) : (
-        <p className='italic '>No active task, click "Generate task" to get one</p>
+        <p className='italic '>
+            {allTasksCompleted
+                ? 'You have completed all tasks! 🎉'
+                : 'No active task, click "Generate task" to get one'}
+        </p>
     );
 
     return (
         <div className='flex flex-col gap-2'>
             <span className='heading-accent-md'>Random Task Generator</span>
-            <div className='w-full px-3 mb-4'>{allTasksCompleted ? renderCompleted : renderTask}</div>
+            <div className='w-full px-3 mb-4'>{noPossibleTasks ? renderNoPossibleTask : renderTask}</div>
             {!allTasksCompleted && (
-                <div className='flex flex-col gap-1 px-3'>
-                    <div className={`flex gap-1 mt ${isLgViewport ? 'w-full' : 'max-w-[420px]'}`}>
-                        <button type='button' className='button-outline w-full' onClick={generateTask}>
-                            {activeTask ? 'Skip task' : 'Generate task'}
-                        </button>
-                        {activeTask && (
-                            <button
-                                type='button'
-                                className='button-outline w-full'
-                                onClick={() => completeTask(activeTask.id)}
-                            >
-                                Complete task
+                <>
+                    <div className='flex flex-col gap-1 px-3'>
+                        <div className={`flex gap-1 mt ${isLgViewport ? 'w-full' : 'max-w-[420px]'}`}>
+                            <button type='button' className='button-outline w-full' onClick={generateTask}>
+                                {activeTask ? 'Skip task' : 'Generate task'}
                             </button>
-                        )}
-                        {/* TODO: ADD configs */}
-                        {/* {isLgViewport && (
-                        <button type='button' className='button-outline w-full'>
-                            Configure task generator
-                        </button>
-                    )} */}
+                            {activeTask && (
+                                <button
+                                    type='button'
+                                    className='button-outline w-full'
+                                    onClick={() => completeTask(activeTask.id)}
+                                >
+                                    Complete task
+                                </button>
+                            )}
+                        </div>
                     </div>
-                    {/* TODO: Add configs */}
-                    {/* {isLgViewport && (
-                    <button type='button' className='button-outline w-full'>
-                        Configure task generator
-                    </button>
-                )} */}
-                </div>
+                    <p className='italic text-xs px-3'>
+                        <span className='material-icons-outlined text-xs'>info</span>Tasks are picked from your table,
+                        according to your current filters
+                    </p>
+                </>
             )}
         </div>
     );
