+++ src/components/TaskTable.js
@@ -1,8 +1,8 @@
 import React, { useMemo } from 'react';
-import _ from 'lodash';
 import { matchSorter } from 'match-sorter';
 import { useSelector } from 'react-redux';
 import tasks from '../data/tasks';
+import ALL_FILTERS from '../util/taskFilters';
 import Cell from './TaskTableCell';
 import Difficulty from './Difficulty';
 import Category from './Category';
@@ -61,13 +61,7 @@ export default function TaskTable() {
         []
     );
     const filters = [
-        difficultyFilter,
-        categoryFilter,
-        subcategoryFilter,
-        skillFilter,
-        completedFilter,
-        todoFilter,
-        ignoredFilter,
+        ...Object.values(ALL_FILTERS)
     ];
     const initialState = isXsViewport ? { hiddenColumns: ['id', 'difficulty', 'category'] } : { hiddenColumns: ['id'] };
     initialState.pageSize = 50;
@@ -90,76 +84,6 @@ export default function TaskTable() {
             enableResizeColumns={!isXsViewport}
         />
     );
-}
-
-function difficultyFilter(record, filterState) {
-    if (filterState.difficulty === null) {
-        return true;
-    }
-    return filterState.difficulty.includes(record.difficulty.label);
-}
-
-function categoryFilter(record, filterState) {
-    if (filterState.categories === null) {
-        return true;
-    }
-    return filterState.categories.includes(record?.category?.label);
-}
-
-function subcategoryFilter(record, filterState) {
-    if (filterState.subcategories === null) {
-        return true;
-    }
-    return filterState.subcategories.includes(record?.subcategory?.label);
-}
-
-function skillFilter(record, filterState, { hiscoresState }) {
-    const taskSkills = record.skillReqs.map(req => req.skill);
-
-    // Check if required skills are included in filter
-    const includeNoReq = filterState.showNoRequirements;
-    const hasValidSkills = _.difference(taskSkills, filterState.skills).length === 0;
-    if (!hasValidSkills || (!includeNoReq && taskSkills.length === 0)) {
-        return false;
-    }
-
-    // Check if level requirements should be ignored
-    if (!hiscoresState || hiscoresState.loading || hiscoresState.error || filterState.showUnmetRequirements) {
-        return true;
-    }
-
-    // Check if level requirements are met
-    let meetsRequirements = true;
-    record.skillReqs.forEach(skillReq => {
-        const hiscores = hiscoresState.skills[skillReq.skill.toLowerCase()];
-        const level = hiscores?.level || 1;
-        meetsRequirements = meetsRequirements && level >= skillReq.level;
-    });
-    return meetsRequirements;
-}
-
-function completedFilter(record, filterState, { tasksState }) {
-    if (filterState.status === 'all') {
-        return true;
-    }
-    const status = !!tasksState[record.id]?.completed;
-    return (filterState.status === 'cmpl') === !!status;
-}
-
-function todoFilter(record, filterState, { tasksState }) {
-    if (filterState.todo === 'all') {
-        return true;
-    }
-    const todo = !!tasksState[record.id]?.todo;
-    return (filterState.todo === 'only') === !!todo;
-}
-
-function ignoredFilter(record, filterState, { tasksState }) {
-    if (filterState.ignored === 'all') {
-        return true;
-    }
-    const ignored = !!tasksState[record.id]?.ignored;
-    return (filterState.ignored === 'only') === !!ignored;
 }
 
 function sortTask(a, b) {
